Revert "Use existing test cases in TestBaseAllocator"
This reverts commit 1e7f21bb61daba8bcaeb3e64311b4fe3a25153d5.
